camcom/requirements.txt
Django==4.1.7
django-extensions==3.1.5 # for shell plus
gunicorn==20.1.0
djangorestframework==3.12.4
sentry-sdk==1.6.0
boto3==1.20.42
psycopg2==2.9
celery==5.2.3
redis==4.1.2
cryptography==41.0.0
django-celery-results==2.4.0
django-storages==1.10.1
django-money==2.1.1
protobuf==3.20.3
tblib==1.7.0 # used for pickling errors when running tests in parallel
uvicorn[standard]==0.18.3
python-memcached==1.59
httpretty==1.1.4
Pillow==9.5.0